:wrench: feat: move enums
