+++ gb28181-common/src/main/java/io/github/lunasaw/gb28181/common/entity/enums/ManufacturerEnum.java
@@ -1,4 +1,4 @@
-package io.github.lunasaw.sip.common.enums;
+package io.github.lunasaw.gb28181.common.entity.enums;
 
 /**
  * @author luna
