+++ gb28181-common/src/main/java/io/github/lunasaw/gb28181/common/entity/enums/TransModeEnum.java
@@ -1,30 +1,29 @@
-package io.github.lunasaw.sip.common.enums;
+package io.github.lunasaw.gb28181.common.entity.enums;
 
 /**
  * @author luna
  * @date 2023/10/13
  */
-public enum StreamModeEnum {
+public enum TransModeEnum {
 
     /**
-     * 数据流传输模式
+     * 传输模式
      */
     UDP("UDP", "UDP"),
-    TCP_ACTIVE("TCP-ACTIVE", "TCP主动"),
-    TCP_PASSIVE("TCP-PASSIVE", "TCP被动"),
+    TCP("TCP", "TCP"),
 
     ;
 
     private final String type;
     private final String desc;
 
-    StreamModeEnum(String type, String desc) {
+    TransModeEnum(String type, String desc) {
         this.type = type;
         this.desc = desc;
     }
 
     public static boolean isValid(String sort) {
-        return UDP.getType().equals(sort) || TCP_ACTIVE.getType().equals(sort) || TCP_PASSIVE.getType().equals(sort);
+        return UDP.getType().equals(sort) || TCP.getType().equals(sort);
     }
 
     public String getType() {
