+++ gb28181-server/src/main/java/io/github/lunasaw/gbproxy/server/entity/InviteEntity.java
@@ -1,8 +1,8 @@
 package io.github.lunasaw.gbproxy.server.entity;
 
 import io.github.lunasaw.gb28181.common.entity.enums.InviteSessionNameEnum;
-import io.github.lunasaw.sip.common.enums.ManufacturerEnum;
-import io.github.lunasaw.sip.common.enums.StreamModeEnum;
+import io.github.lunasaw.gb28181.common.entity.enums.ManufacturerEnum;
+import io.github.lunasaw.gb28181.common.entity.enums.StreamModeEnum;
 import org.apache.commons.lang3.StringUtils;
 
 /**
